Move tests to a new api folder
